requirements.txt
wandb==0.18.7
apify-client==1.6.1
bittensor==9.9.0
jupyter==1.0.0
numpy>=2.0.1
pydantic==2.10.6
python-dotenv==1.0.0
pytz==2023.3.post1
rich==13.7.0
pandas~=2.2.2
cryptography==45.0.4
requests==2.32.4
huggingface-hub==0.27.1
datasets~=2.20.0
pyarrow>=17.0.0
fsspec==2024.5.0
psutil==5.9.8
loguru==0.7.3
duckdb==1.3.2
langcodes==3.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
httpx==0.28.1
curl-cffi==0.7.3
beautifulsoup4==4.12.3
scrapingbee==2.0.2

# RESI only handles real estate data - social media dependencies removed:
# asyncpraw==7.8.0                    # Reddit scraper
# google-api-python-client==2.167.0   # YouTube scraper
# youtube-transcript-api==1.2.0       # YouTube scraper
# isodate==0.7.2                     # YouTube scraper